Switched integer and pointer literals to long long type
This way we ensure that those are at least 64 bits on
any platform.

+++ include/smack/BoogieAst.h
@@ -32,11 +32,11 @@ public:
   static const Expr* impl(const Expr* l, const Expr* r);
   static const Expr* lit(bool b);
   static const Expr* lit(std::string v);
-  static const Expr* lit(unsigned v) { return lit((unsigned long) v); }
-  static const Expr* lit(unsigned long v);
-  static const Expr* lit(long v);
+  static const Expr* lit(unsigned v) { return lit((unsigned long long) v); }
+  static const Expr* lit(unsigned long long v);
+  static const Expr* lit(long long v);
   static const Expr* lit(std::string v, unsigned w);
-  static const Expr* lit(unsigned long v, unsigned w);
+  static const Expr* lit(unsigned long long v, unsigned w);
   static const Expr* lit(bool n, std::string s, std::string e, unsigned ss, unsigned es);
   static const Expr* lit(RModeKind v);
   static const Expr* neq(const Expr* l, const Expr* r);
@@ -97,12 +97,12 @@ class IntLit : public Expr {
   std::string val;
 public:
   IntLit(std::string v) : val(v) {}
-  IntLit(unsigned long v) {
+  IntLit(unsigned long long v) {
     std::stringstream s;
     s << v;
     val = s.str();
   }
-  IntLit(long v) {
+  IntLit(long long v) {
     std::stringstream s;
     s << v;
     val = s.str();
@@ -115,7 +115,7 @@ class BvLit : public Expr {
   unsigned width;
 public:
   BvLit(std::string v, unsigned w) : val(v), width(w) {}
-  BvLit(unsigned long v, unsigned w) : width(w) {
+  BvLit(unsigned long long v, unsigned w) : width(w) {
     std::stringstream s;
     s << v;
     val = s.str();
